requirements.txt
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
click>=8.1.0
